scripts/requirements.txt
# Beyond the Code Content Pipeline Dependencies

# Environment configuration (for local development)
python-dotenv>=1.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
trafilatura>=1.6.0  # For article extraction (used by hn_scraper.py)

# RSS parsing
feedparser>=6.0.0

# Configuration
pyyaml>=6.0

# LLM Providers
anthropic>=0.18.0
google-genai>=1.0.0
groq>=0.4.0
openai>=1.0.0
